requirements.txt
Flask
openai
python-dotenv
whisper
scipy
sounddevice
